+++ Sources/Sentry/SentryTraceContext.m
@@ -51,9 +51,16 @@ NS_ASSUME_NONNULL_BEGIN
         return nil;
 
     NSString *userSegment;
-    if (scope.userObject.data[@"segment"] &&
-        [scope.userObject.data[@"segment"] isKindOfClass:[NSString class]])
-        userSegment = scope.userObject.data[@"segment"];
+
+    if (scope.userObject.segment) {
+        userSegment = scope.userObject.segment;
+    } else {
+        // For backwards compatibility
+        if (scope.userObject.data[@"segment"] &&
+            [scope.userObject.data[@"segment"] isKindOfClass:[NSString class]]) {
+            userSegment = scope.userObject.data[@"segment"];
+        }
+    }
 
     NSString *sampleRate = nil;
     if ([tracer.context isKindOfClass:[SentryTransactionContext class]]) {